Fix build where "git describe" fails
This was most likely due to Travis CI uses a shallow clone of the
repository and fetches only the 50 latest commits. It was most likely
more than 50 commits since the last tag cause "git describe" to fail.
Adding a before script step in the .travis.yml file which does
"git fetch --unshallow" seems to fix the problem.