Add early return on null pointer
This patch adds a reduced test case for a C macro defining a function
call `(bar)(a)`, i.e., calling the function `bar` on `a`. The extra
parentheses make the test result in a segmentation fault due to null
pointer dereferencing. To fix the problem the null pointer check is
performed as part of an early return. This makes the code more robust
and allows removing a later null check.