Fix anonymous composites in clang 19.1.7
In some new clang versions, the name (cursor.spelling) of anonymous structs,
unions and enums was set to an informative message instead of an empty
string. I.e. "enum (unnamed at [declaration_file.h:48]"

This fixes the errors that happen due to that in a backwards compatible
way, since it still checks for empty spellings.

This patch is possibly overeager in checking the cursor.spellings for
these new special values, I haven't checked at every single location
whether the check is necessary.